+++ src/main/java/io/github/cottonmc/templates/model/RetexturingBakedModel.java
@@ -6,7 +6,6 @@ import net.fabricmc.fabric.api.renderer.v1.mesh.Mesh;
 import net.fabricmc.fabric.api.renderer.v1.mesh.MutableQuadView;
 import net.fabricmc.fabric.api.renderer.v1.model.ForwardingBakedModel;
 import net.fabricmc.fabric.api.renderer.v1.render.RenderContext;
-import net.fabricmc.fabric.api.rendering.data.v1.RenderAttachedBlockView;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.client.MinecraftClient;
@@ -62,7 +61,8 @@ public abstract class RetexturingBakedModel extends ForwardingBakedModel {
 	
 	@Override
 	public void emitBlockQuads(BlockRenderView blockView, BlockState state, BlockPos pos, Supplier<Random> randomSupplier, RenderContext context) {
-		BlockState theme = (((RenderAttachedBlockView) blockView).getBlockEntityRenderAttachment(pos) instanceof BlockState s) ? s : null;
+		//FabricBlockView itf-inject
+		BlockState theme = blockView.getBlockEntityRenderData(pos) instanceof BlockState s ? s : null;
 		if(theme == null || theme.isAir()) {
 			getUntintedRetexturedMesh(new CacheKey(state, tam.getDefaultAppearance())).outputTo(context.getEmitter());
 			return;